// lib/wordpress.d.ts
interface RenderedContent {
  rendered: string;
  protected: boolean;
}
// ...
export interface Project {
  id: number;
  slug: string;
  title: RenderedContent;
  excerpt: RenderedContent;
  content: RenderedContent;
  featured_media_url: string;
  featured_media_width: number;
  featured_media_height: number;
  categories: number[];
  categoryNames: string[];
  _embedded?: {
    'wp:featuredmedia'?: any[];
  };
}

interface ProjectsGridProps {
  initialProjects: Project[];
  categories: Category[];
}

// Taxonomy types
interface Taxonomy {
  id: number;
  count: number;
  description: string;
  link: string;
  name: string;
  slug: string;
  meta: Record<string, unknown>;
}

export interface Category {
  id: number;
  count: number;
  description: string;
  link: string;
  name: string;
  slug: string;
  taxonomy: string;
  parent: number;
}